added FindUnit Function
added FindUnit Function which when called selects the best unit from the board based on current player, and where on the board the player selected. this prevents the player from selecting and moving more than one of their units, as well as prevents them from selecting one of their opponent's units.